Add ticktock multiqueue
the ticktock multiqueue found in FixedDeadliner exploits a bunch of quirks of our env to ensure that we get a lockless consumer with multiple producers using just atomic spins. it supports 2 simultaneous producers and in almost all cases, is both blazing fast and has superb memory locality.

it also solves a crash bug, I guess.